【AC】[Kyuride Kagamiz Programming Contest (Easy)]D - 虫歯(Cavity)

+++ k2pc-easy/k2pc001_e4.cpp
@@ -33,7 +33,7 @@ std::istream &operator>>(std::istream &in, queue<int> &o) {
     return in;
 }
 
-bool contain(set<ll> &s, int a) { return s.find(a) != s.end(); }
+bool contain(set<ll> &s, ll a) { return s.find(a) != s.end(); }
 
 typedef priority_queue<ll, vector<ll>, greater<ll> > PQ_ASK;
